Readability on index loop, moved examples into example directory
Thanks for the ideas @tycrek
