+++ package.json
@@ -21,7 +21,9 @@
     "@types/express": "^4.17.13",
     "@types/pug": "^2.0.5",
     "dotenv": "^10.0.0",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "ts": "^0.2.2",
+    "typescript": "^4.3.5"
   },
   "devDependencies": {
     "pug": "^3.0.2"
