+++ src/core/pages.ts
@@ -1,13 +1,13 @@
 import { PageType } from "./types";
 
 // Index
-import index from "../pages/home";
+import index from "../pages/examples/home";
 export const IndexPage: PageType = index;
 
 // About
-import about from "../pages/about";
+import about from "../pages/examples/about";
 export const AboutPage: PageType = about;
 
 // API example
-import api_example from "../pages/api_example";
+import api_example from "../pages/examples/api_example";
 export const API_Example: PageType = api_example;
